added function in util to check if a tx is confirmed. Transaction confirmation is to unpredictable to use a time.sleep so this function will wait for the max time before we say that the transaction will never be confirmed.
Changed the way sawtooth permissions work. Previously when a peer joined a a committee it's validator key needed to be added to the list of
administrators of the network (so it could add/remove members). Now each container shares one admin key via a class static variables that is
set when the first peer make a genesis block for the first time. All other peers in all other committees share this key.

Made instance_a and instance_b publicly visible (needed for testing).

Ideal view change timeout and commit view changed changed to 30sec and 15 sec respectively.

update membership was split into two functions one that does logging and parameter checking (update_committee) and another that does the
update (__update_on_chain_settings) this may submit an update twice if it is not confirmed soon enough the first time (i.e it may add one
or two blocks)

Improved tolerance of blocks method in SawtoothPBFT.py so that if it cant get a response from the validator it stills returns 0 blocks.

all tests have been adjusted per these changes
committee churn is skipped because there is some unknown issue with it (intermediate failure). This is considered outside the scope of
improving committee size.